Fix Pushable Object rotation
In some uncommon cases, if the player kept the Left or Right key pressed during the alignment process to grab a pushable, it was sometimes possible for Lara to keep rotating beyond the alignment goal, allowing the player to move the block in cardinal directions that bypassed the collision checks.

These two lines force the code to ignore any stored Turning Rate value when the alignment process starts.